musicapppr_backend/AI/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
transformers
torch
Pillow
numpy
opencv-python
fer